Move Edges to component file
