+++ src/constants.ts
@@ -1,0 +1,2 @@
+export const height: number = 800;
+export const width: number = 1200;
